# common/swaggers/data/swagger_output_jsdoc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tabcolab",
    "version": "1.0.0",
    "description": "Your API Description"
  },
  "servers": [
    {
      "url": "http://localhost:5050/api/1.0"
    }
  ],
...
  "paths": {
...
    "/groups/items/search": {
      "get": {
        "tags": [
          "Items"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Search items in all groups by keyword and filter by item_type",
        "description": "Retrieve items in all groups that match the provided keyword.<br><br>Here's the API format :<br>`/groups/items/search?keyword={keyword}&itemTypes={item_type1},{item_type2}`<br>Where keyword is a required field. <br>If itemTypes is not specified or empty, it defaults to returning matching results for all the item_types.<br><br>Below are potential scenarios for the keyword parameter:<br> 1. A single space is utilized to separate two keywords, as seen in \"keywordA keywordB.\" This format triggers separate searches for each keyword.<br>2. If the parameter consists solely of spaces, like \" \", the search will include results containing empty strings and spaces.<br>3. Consecutive spaces within the parameter will trigger additional searches for empty strings. For instance, in \"keywordC&nbsp;&nbsp;keywordD,\" the search will include keywordC, keywordD, as well as empty strings and spaces.",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "description": "The keyword to search for items.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "itemTypes",
            "description": "Filter by selected item types.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Item/properties/item_type"
              }
            },
            "style": "form",
            "explode": false
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "allOf": [
                      {
                        "type": "object",
                        "properties": {
                          "groupId": {
                            "type": "string",
                            "description": "The ID of the group"
                          }
                        }
                      },
                      {
                        "$ref": "#/components/schemas/Item"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Query Parameters"
          },
          "401": {
            "description": "Invalid JWT or Missing JWT token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/groups/{group_id}/items/{item_id}": {
      "parameters": [
        {
          "name": "group_id",
          "in": "path",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/Group/properties/group_id"
          },
          "example": 1
        },
        {
          "name": "item_id",
          "in": "path",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/Item/properties/item_id"
          },
          "example": 1
        }
      ],
      "patch": {
        "tags": [
          "Items"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Move item (tab, note, todo) within or between existing Groups",
        "description": "Move item (tab, note, todo) :<br> 1. Within existing Groups <br> 2. Between existing Groups",
        "operationId": "groupItemDragPatchGroup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/requestBodies/ItemMoveinGroup"
                  },
                  {
                    "$ref": "#/components/requestBodies/ItemMovetoGroup"
                  }
                ]
              },
              "examples": {
                "ItemMoveinGroup": {
                  "summary": "Moving an item within the same group.",
                  "value": {
                    "targetItem_position": 2
                  }
                },
                "ItemMovetoGroup": {
                  "summary": "Moving an item to a different group.",
                  "value": {
                    "targetItem_position": 1,
                    "targetGroup_id": "2"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Invalid JWT or Missing JWT token"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "example": 2
          },
          {
            "name": "item_id",
            "in": "path",
            "example": 6
          }
        ],
        "tags": [
          "Items"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Remove item (tab, note, todo) from existing Group",
        "operationId": "groupPatchItem",
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Invalid JWT or Missing JWT token"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
  "components": {
...
    "requestBodies": {
...
      "ItemMoveinGroup": {
        "description": "Move item within existing Groups",
        "type": "object",
        "properties": {
          "targetItem_position": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "ItemMovetoGroup": {
        "description": "Move item between existing Groups",
        "type": "object",
        "properties": {
          "targetItem_position": {
            "type": "integer"
          },
          "targetGroup_id": {
            "$ref": "#/components/schemas/Group/properties/group_id"
          }
        }
      }
    },
    "schemas": {
...
      "Item": {
        "type": "object",
        "properties": {
          "item_id": {
            "type": "string"
          },
          "item_type": {
            "type": "integer",
            "enum": [
              0,
              1,
              2
            ],
            "description": "0: Tab, 1: Note, 2: Todo"
          }
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/Tab"
          },
          {
            "$ref": "#/components/schemas/Note"
          },
          {
            "$ref": "#/components/schemas/Todo"
          }
        ]
      },
      "Tab": {
        "type": "object",
        "properties": {
          "browserTab_favIconURL": {
            "type": "string"
          },
          "browserTab_title": {
            "type": "string"
          },
          "browserTab_url": {
            "type": "string"
          },
          "browserTab_id": {
            "type": "integer"
          },
          "browserTab_index": {
            "type": "integer"
          },
          "browserTab_active": {
            "type": "boolean"
          },
          "browserTab_status": {
            "type": "string",
            "enum": [
              "complete",
              "loading",
              "unloaded"
            ]
          },
          "windowId": {
            "type": "integer",
            "example": 1348438505
          },
          "note_content": {
            "type": "string"
          },
          "note_bgColor": {
            "type": "string"
          }
        }
      },
      "Note": {
        "type": "object",
        "properties": {
          "note_content": {
            "type": "string"
          },
          "note_bgColor": {
            "type": "string"
          }
        }
      },
      "Todo": {
        "type": "object",
        "properties": {
          "note_content": {
            "type": "string"
          },
          "note_bgColor": {
            "type": "string"
          },
          "doneStatus": {
            "type": "boolean",
            "description": "Indicates whether the todo is done or not"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT (JSON Web Token) authentication in the request header.\n```\nAuthorization: Bearer <token>\n```\nKeyin your JWT `<token>` below:\n`eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiMSIsImVtYWlsIjoidXNlckBleGFtcGxlLmNvbSIsInZlcnNpb24iOiJ2MS4wIiwiaWF0IjoxNzExOTA1NTc4LCJleHAiOjE3MTI1MTAzNzh9.umagmr0uU8ItdxPGz6JSTGcRXiitFdTGCCfzHCRB4Ew`\n"
      }
    }
  }
}